refactor: only test 10, 100, 1000
