+++ packages/store-sync/src/storageAdapter10.bench.ts
@@ -1,11 +1,11 @@
 import { bench, describe } from "vitest";
 import { createRecsStorage, createSqliteStorageAdapter, createZustandStorageAdapter } from "../test/utils";
 import { logsToBlocks } from "../test/logsToBlocks";
-import worldRpcLogs from "../../../test-data/world-logs-1000.json";
+import worldRpcLogs from "../../../test-data/world-logs-10.json";
 
 const blocks = logsToBlocks(worldRpcLogs);
 
-describe("Storage Adapter 1000 logs", () => {
+describe("Storage Adapter 10 logs", () => {
   bench("recs: `storageAdapter`", async () => {
     const { storageAdapter } = createRecsStorage();
